modularize out metrics reading/writing, migrate `/sugois` to use said module
